[Frontend] Move APIServerProcessManager target server fn (#38115)
