+++ tests/entrypoints/test_api_server_process_manager.py
@@ -89,9 +89,7 @@ def test_api_server_process_manager_init(api_server_args, with_stats_update):
             assert not proc.is_alive()
 
 
-@patch(
-    "vllm.entrypoints.cli.serve.run_api_server_worker_proc", mock_run_api_server_worker
-)
+@patch("vllm.v1.utils.run_api_server_worker_proc", mock_run_api_server_worker)
 def test_wait_for_completion_or_failure(api_server_args):
     """Test that wait_for_completion_or_failure works with failures."""
     global WORKER_RUNTIME_SECONDS
